% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAGSWrapper.R
\name{JAGS}
\alias{JAGS}
\title{JAGS}
\usage{
JAGS(
  data = NULL,
  version = "1",
  actualExporter = FALSE,
  aggregatedChains = TRUE,
  autoCorPlot = FALSE,
  autoCorPlotLags = 20,
  autoCorPlotType = "lines",
  bivariateScatterDiagonalType = "density",
  bivariateScatterOffDiagonalType = "hexagon",
  bivariateScatterPlot = FALSE,
  burnin = 500,
  chains = 3,
  colorScheme = "colorblind",
  customInference = list(list(ciLevel = 0.95, dataSplit = list(types = "unknown", value =
    ""), ess = TRUE, hdiLevel = 0.95, inferenceCi = FALSE, inferenceCiLevel = 0.95,
    inferenceCustomHigh = 1, inferenceCustomLow = 0, inferenceData = list(types =
    "unknown", value = ""), inferenceHdi = FALSE, inferenceHdiLevel = 0.95,
    inferenceManual = FALSE, mean = TRUE, median = TRUE, mode = TRUE, name = "Plot 1",
    overlayGeomType = "density", overlayHistogramBinWidthType = "sturges",
    overlayHistogramManualNumberOfBins = 30, parameter = NULL, 
     parameterOrder =
    "orderMean", parameterSubset = "", plotCustomHigh = 1, plotCustomLow = 0,
    plotInterval = "ci", plotsType = "", rhat = TRUE, savageDickey = FALSE,
    savageDickeyPoint = 0, savageDickeyPosteriorMethod = "samplingPosteriorPoint",
    savageDickeyPosteriorSamplingType = "normalKernel", savageDickeyPriorHeight = 0,
    savageDickeyPriorMethod = "sampling", savageDickeySamplingType = "normalKernel", sd =
    TRUE, shadeIntervalInPlot = FALSE)),
  densityPlot = FALSE,
  deviance = FALSE,
  exportSamplesFile = "",
  histogramPlot = FALSE,
  initialValues = list(list(levels = list(), name = "Parameter", values = list()),
    list(levels = list(), name = "R Code", values = list())),
  legend = TRUE,
  model = list(columns = list(), model = "model{\\n\\n}", modelOriginal =
    "model{\\n\\n}", parameters = list()),
  monitoredParameters = list(types = list(), value = list()),
  monitoredParametersShown = list(types = list(), value = list()),
  plotHeight = 320,
  plotWidth = 480,
  resultsFor = "allParameters",
  samples = 2000,
  seed = 1,
  setSeed = FALSE,
  thinning = 1,
  tracePlot = FALSE,
  userData = list(list(levels = list(), name = "Parameter", values = list()), list(levels
    = list(), name = "R Code", values = list()))
)
}
\description{
JAGS
}
